Add and change some features
I've added a --show option to replace --attachments, --location, and
--notes. These are now "--show attachments,location,notes" (assuming you
want all three). I've also added "free" and "busy" (which mean the same
thing) to the items --show recognizes. They add a free/busy field before
the start time. There's also a new "day" item for --show that prepends
the start and end time with the day of week.

I've also added a --free-days option that will output only a list of
non-busy days in the given interval.